+++ src/SlackConnector.Tests.Unit/SlackConnectorTests/ConnectedStatusTests.cs
@@ -7,6 +7,7 @@ using Should;
 using SlackConnector.Connections;
 using SlackConnector.Connections.Handshaking;
 using SlackConnector.Connections.Models;
+using SlackConnector.Connections.Responses;
 using SlackConnector.Connections.Sockets;
 using SlackConnector.Exceptions;
 using SlackConnector.Models;
