+++ src/SlackConnector.Tests.Unit/SlackConnectorTests/HubsTests.cs
@@ -5,6 +5,7 @@ using Should;
 using SlackConnector.Connections;
 using SlackConnector.Connections.Handshaking;
 using SlackConnector.Connections.Models;
+using SlackConnector.Connections.Responses;
 using SlackConnector.Connections.Sockets;
 using SlackConnector.Tests.Unit.Stubs;
 using SpecsFor;
