+++ src/SlackConnector/Connections/Messaging/ChannelMessenger.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using Newtonsoft.Json;
 using RestSharp;
 using SlackConnector.Connections.Models;
+using SlackConnector.Connections.Responses;
 using SlackConnector.Exceptions;
 
 namespace SlackConnector.Connections.Messaging
