+++ src/SlackConnector/Connections/Messaging/ChatMessenger.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Newtonsoft.Json;
 using RestSharp;
 using SlackConnector.Connections.Models;
+using SlackConnector.Connections.Responses;
 using SlackConnector.Exceptions;
 using SlackConnector.Models;
