+++ src/SlackConnector/Connections/Responses/JoinChannelResponse.cs
@@ -1,4 +1,6 @@
-﻿namespace SlackConnector.Connections.Models
+﻿using SlackConnector.Connections.Models;
+
+namespace SlackConnector.Connections.Responses
 {
     internal class JoinChannelResponse : StandardResponse
     {
